Step1: refactor geotools project and build changes for the same
