+++ org.geotools/.project
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <projectDescription>
-	<name>org.geotools</name>
+	<name>org.mwc.cmap.geotools</name>
 	<comment></comment>
 	<projects>
 	</projects>
