+++ org.geotools/src/org/mwc/cmap/geotools/gt2plot/GeoToolsPainter.java
@@ -12,7 +12,7 @@
  *    but WITHOUT ANY WARRANTY; without even the implied warranty of
  *    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. 
  */
-package org.mwc.cmap.gt2plot.proj;
+package org.mwc.cmap.geotools.gt2plot;
 
 import java.awt.Color;
 import java.awt.Graphics2D;
