+++ org.mwc.asset.NetAsset2/META-INF/MANIFEST.MF
@@ -14,7 +14,8 @@ Require-Bundle: org.eclipse.ui,
  org.mwc.cmap.plotViewer,
  org.mwc.debrief.legacy,
  org.mwc.debrief.core,
- org.mwc.cmap.gt2Plot;bundle-version="1.0.0"
+ org.mwc.cmap.gt2Plot;bundle-version="1.0.0",
+ org.mwc.cmap.geotools;bundle-version="1.0.7"
 Bundle-ActivationPolicy: lazy
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Bundle-ClassPath: .,
