+++ org.mwc.cmap.NaturalEarth/META-INF/MANIFEST.MF
@@ -7,7 +7,7 @@ Bundle-Activator: org.mwc.cmap.naturalearth.Activator
 Bundle-Vendor: planetmayo
 Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime,
- org.geotools,
+ org.mwc.cmap.geotools,
  org.mwc.cmap.legacy;bundle-version="1.0.190",
  org.mwc.cmap.gt2Plot,
  org.mwc.cmap.core
