+++ org.mwc.cmap.combined.feature/feature.xml
@@ -292,13 +292,6 @@ Library.
          unpack="false"/>
 
    <plugin
-         id="org.geotools"
-         download-size="0"
-         install-size="0"
-         version="0.0.0"
-         unpack="false"/>
-
-   <plugin
          id="org.eclipse.nebula.cwt"
          download-size="0"
          install-size="0"
@@ -327,6 +320,13 @@ Library.
 
    <plugin
          id="org.mwc.cmap.media"
+         download-size="0"
+         install-size="0"
+         version="0.0.0"
+         unpack="false"/>
+
+   <plugin
+         id="org.mwc.cmap.geotools"
          download-size="0"
          install-size="0"
          version="0.0.0"
