+++ org.mwc.cmap.gt2Plot/META-INF/MANIFEST.MF
@@ -8,11 +8,10 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.core.runtime,
  org.mwc.cmap.core;bundle-version="1.0.143",
  org.mwc.cmap.legacy;bundle-version="1.0.137",
- org.geotools
+ org.mwc.cmap.geotools
 Bundle-RequiredExecutionEnvironment: JavaSE-1.7
 Bundle-ClassPath: .
 Export-Package: org.mwc.cmap.gt2plot,
- org.mwc.cmap.gt2plot.data,
- org.mwc.cmap.gt2plot.proj,
+ org.mwc.cmap.gt2plot.calc,
  org.mwc.cmap.gt2plot.views
 Bundle-ActivationPolicy: lazy
