+++ org.mwc.cmap.plotViewer/src/org/mwc/cmap/plotViewer/editors/CorePlotEditor.java
@@ -84,10 +84,10 @@ import org.mwc.cmap.core.interfaces.IResourceProvider;
 import org.mwc.cmap.core.operations.IUndoable;
 import org.mwc.cmap.core.property_support.EditableWrapper;
 import org.mwc.cmap.core.ui_support.PartMonitor;
-import org.mwc.cmap.gt2plot.data.GeoToolsLayer;
-import org.mwc.cmap.gt2plot.data.ShapeFileLayer;
-import org.mwc.cmap.gt2plot.data.WorldImageLayer;
-import org.mwc.cmap.gt2plot.proj.GtProjection;
+import org.mwc.cmap.geotools.gt2plot.GeoToolsLayer;
+import org.mwc.cmap.geotools.gt2plot.GtProjection;
+import org.mwc.cmap.geotools.gt2plot.ShapeFileLayer;
+import org.mwc.cmap.geotools.gt2plot.WorldImageLayer;
 import org.mwc.cmap.naturalearth.Activator;
 import org.mwc.cmap.naturalearth.wrapper.NELayer;
 import org.mwc.cmap.plotViewer.PlotViewerPlugin;
