+++ org.mwc.debrief.core/META-INF/MANIFEST.MF
@@ -66,5 +66,6 @@ Require-Bundle: org.mwc.cmap.core,
  org.eclipse.jface,
  org.eclipse.ui.workbench,
  org.eclipse.nebula.jface.cdatetime;bundle-version="0.14.0",
- org.eclipse.nebula.widgets.cdatetime;bundle-version="0.14.0"
+ org.eclipse.nebula.widgets.cdatetime;bundle-version="0.14.0",
+ org.mwc.cmap.geotools;bundle-version="1.0.7"
 Bundle-ActivationPolicy: lazy
