+++ org.mwc.debrief.lite/conf/manifest.txt
@@ -4,7 +4,7 @@ Created-By: 1.8.0_211-b12 (Oracle Corporation)
 Main-Class: org.mwc.debrief.lite.DebriefLiteApp
 SplashScreen-Image: images/splash.gif
 Implementation-Title: Debrief Lite
-Implementation-Version: 1.0.396
+Implementation-Version: 1.0.397
 Implementation-Vendor: DeepBlue C Ltd
 Class-Path: . lib/cmap_legacy.jar lib/debrief_legacy.jar lib/jt-zonals
  tats-1.3.1.jar lib/imageio-ext-tiff-1.1.10.jar lib/imageio-ext-utilit
