+++ org.mwc.debrief.satc.core/META-INF/MANIFEST.MF
@@ -14,7 +14,7 @@ Require-Bundle: org.eclipse.ui,
  org.eclipse.nebula.jface.cdatetime,
  org.junit,
  org.eclipse.help;bundle-version="3.6.0",
- org.geotools,
+ org.mwc.cmap.geotools,
  org.mwc.cmap.jfreechart,
  org.mwc.cmap.legacy,
  org.mwc.debrief.track_shift;bundle-version="1.0.72"
